+++ packages/spotlight/SpotlightRootDecorator/SpotlightRootDecorator.js
@@ -7,6 +7,7 @@
  */
 
 import hoc from '@enact/core/hoc';
+import {is} from '@enact/core/keymap';
 import React from 'react';
 
 import Spotlight from '../src/spotlight';
@@ -107,9 +108,15 @@ const SpotlightRootDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		};
 
-		handleKeyDown = () => {
-			this.containerRef.current.classList.add('non-touch-mode');
-			this.containerRef.current.classList.remove('touch-mode');
+		handleKeyDown = (ev) => {
+			const {keyCode} = ev;
+			if (is('enter', keyCode) && this.containerRef.current.classList.contains('touch-mode')) {
+				// Prevent onclick event trigger by enter key
+				ev.preventDefault();
+			} else {
+				this.containerRef.current.classList.add('non-touch-mode');
+				this.containerRef.current.classList.remove('touch-mode');
+			}
 		};
 
 		navigableFilter = (elem) => {